+++ src/controller/RecipeController.js
@@ -1,17 +1,3 @@
-const multer = require("multer");
-const upload = multer({
-    storage: multer.diskStorage({
-        destination(req, file, done) {
-            done(null, 'public/img/');
-        },
-        filename(req, file, done) {
-            const ext = path.extname(file.originalname);
-            done(null, path.basename(file.originalname, ext) + ext);
-        }
-    }),
-    limits: {fileSize : 5*1024*1024},
-})
-
 const models = require("../model");
 
 exports.main = (req, res) => {
