+++ src/views/index.ejs
@@ -1,6 +1,6 @@
 <html>
     <head>
-        <title>RECIPE of GOD에 오신 것을 환영합니다</title>
+        <title>신메뉴</title>
 
         <meta charset="utf-8">
